requirements.txt
imagehash
scikit-image
Pillow
pandas
requests
cairosvg
numpy


